# add extra guard on connection management for 1D fabric kernel